Issue #2110: minor cleanup
