+++ app/src/main/java/org/mozilla/tv/firefox/navigationoverlay/channels/ChannelTile.kt
@@ -2,13 +2,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.channel
+package org.mozilla.tv.firefox.navigationoverlay.channels
 
 import android.widget.ImageView
 
-/**
- * TODO
- */
 data class ChannelTile(
         val url: String,
         val title: String,
