+++ app/src/main/res/layout/default_channel.xml
@@ -1,18 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <TextView
         android:id="@+id/channelTitle"
+        style="@style/TextAppearance.AppCompat.Medium"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        style="@style/TextAppearance.AppCompat.Medium"
         android:layout_marginStart="12dp"
         android:layout_marginTop="24dp"
         android:layout_marginBottom="24dp"
-        android:text="@string/menu_settings"
-        android:fontFamily="@string/font_ember_regular"/>
+        android:fontFamily="@string/font_ember_regular"
+        android:text="@string/menu_settings" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/channelTileContainer"
